+++ doc/eurlexpkg.R
@@ -17,26 +17,35 @@ results <- dirs %>% select(-force,-date)
 
 ## -----------------------------------------------------------------------------
 query_dir %>% 
-  glue::as_glue() # for nicer printing
+  cat() # for nicer printing
 
 elx_make_query(resource_type = "caselaw") %>% 
-  glue::as_glue()
+  cat()
 
 elx_make_query(resource_type = "manual", manual_type = "SWD") %>% 
-  glue::as_glue()
+  cat()
 
 
 ## -----------------------------------------------------------------------------
 elx_make_query(resource_type = "directive", include_date = TRUE, include_force = TRUE) %>% 
-  glue::as_glue()
+  cat()
 
 # minimal query: elx_make_query(resource_type = "directive")
 
 elx_make_query(resource_type = "recommendation", include_date = TRUE, include_lbs = TRUE) %>% 
-  glue::as_glue()
+  cat()
 
 # minimal query: elx_make_query(resource_type = "recommendation")
 
+
+## -----------------------------------------------------------------------------
+# request documents from directory 18 ("Common Foreign and Security Policy")
+# and sector 3 ("Legal acts")
+
+elx_make_query(resource_type = "any",
+               directory = "18",
+               sector = 3) %>% 
+  cat()
 
 ## ----runquery, eval=FALSE-----------------------------------------------------
 #  results <- elx_run_query(query = query_dir)
@@ -65,17 +74,13 @@ rec_eurovoc %>%
 
 
 ## ----eurovoctable-------------------------------------------------------------
-
 eurovoc_lookup <- elx_label_eurovoc(uri_eurovoc = rec_eurovoc$eurovoc)
 
 print(eurovoc_lookup)
 
-
 ## ----appendlabs---------------------------------------------------------------
-
 rec_eurovoc %>% 
   left_join(eurovoc_lookup)
-
 
 ## -----------------------------------------------------------------------------
 eurovoc_lookup <- elx_label_eurovoc(uri_eurovoc = rec_eurovoc$eurovoc,
@@ -85,7 +90,6 @@ eurovoc_lookup <- elx_label_eurovoc(uri_eurovoc = rec_eurovoc$eurovoc,
 rec_eurovoc %>% 
   left_join(eurovoc_lookup) %>% 
   select(celex, eurovoc, labels)
-
 
 ## ----getdatapur, message = FALSE, warning=FALSE, error=FALSE------------------
 # the function is not vectorized by default
@@ -117,7 +121,9 @@ dirs %>%
 
 ## -----------------------------------------------------------------------------
 dirs %>% 
-  ggplot(aes(x = as.Date(date), y = celex)) +
+  filter(!is.na(force)) %>% 
+  mutate(date = as.Date(date)) %>% 
+  ggplot(aes(x = date, y = celex)) +
   geom_point(aes(color = force), alpha = 0.1) +
   theme(axis.text.y = element_blank(),
         axis.line.y = element_blank(),